전남대_18조_가든_9주차

+++ src/main/java/com/example/demo/mentoring/MentorPostService.java
@@ -1,10 +1,11 @@
 package com.example.demo.mentoring;
 
-import com.example.demo.config.errors.exception.Exception500;
-import com.example.demo.config.errors.exception.Exception400;
+import com.example.demo.config.errors.exception.Exception401;
+import com.example.demo.config.errors.exception.Exception403;
 import com.example.demo.config.errors.exception.Exception404;
 import com.example.demo.mentoring.contact.ContactJPARepository;
 import com.example.demo.mentoring.contact.NotConnectedRegisterUser;
+import com.example.demo.user.Role;
 import com.example.demo.user.User;
 import com.example.demo.user.userInterest.UserInterest;
 import com.example.demo.user.userInterest.UserInterestJPARepository;
@@ -14,11 +15,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.example.demo.config.errors.exception.Exception400;
-import com.example.demo.config.errors.exception.Exception404;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
+import java.util.Locale;
 import java.util.stream.Collectors;
 
 @Transactional
@@ -31,39 +32,41 @@ public class MentorPostService {
 
     //mentorPost생성
     @Transactional
-    public void createMentorPost(MentorPostRequest.CreateDTO createDTO, User writer) {
-        MentorPost mentorPost = new MentorPost( writer, createDTO.getTitle(), createDTO.getContent());
-        try {
-            mentorPostJPARepository.save(mentorPost);
-        } catch (Exception e) {
-            throw new Exception500("unknown server error");
+    public MentorPostResponse.MentorPostIdDTO createMentorPost(MentorPostRequest.CreateMentorPostDTO createMentorPostDTO, User writer) {
+        isMentor(writer);
+
+        Pageable recentPostTime = PageRequest.of(0, 1);
+        Page<LocalDateTime> recentPostDate = mentorPostJPARepository.createdRecentPost(writer.getId(), recentPostTime);
+        if(recentPostDate != null && !recentPostDate.getContent().isEmpty())
+        {
+            Duration duration = Duration.between(recentPostDate.getContent().get(0), LocalDateTime.now());
+            if (duration.toMinutes() < 5) {
+                throw new Exception403("5분이내로 글을 작성 할 수 없습니다.");
+            }
         }
+
+        MentorPost mentorPost = new MentorPost( writer, createMentorPostDTO.getTitle(), createMentorPostDTO.getContent());
+
+        mentorPostJPARepository.save(mentorPost);
+
+        return new MentorPostResponse.MentorPostIdDTO(mentorPost);
     }
 
    /* 1. mentorPostList를 조회
     2. 각 List당 writer별 writerInterests를 조회
     3. MentorPostDTO 생성*/
-    public List<MentorPostResponse.MentorPostAllDTO> findAllMentorPost(int page) {
+    public List<MentorPostResponse.MentorPostAllDTO> findAllMentorPost(String search, String keyword, int page) {
         Pageable pageable = PageRequest.of(page,5);
 
-        Page<MentorPost> pageContent = mentorPostJPARepository.findAll(pageable);
+        MentorSearchCategory mentorSearchCategory = MentorSearchCategory.valueOf(search.toUpperCase(Locale.ROOT));
 
-        if(pageContent.getTotalPages() == 0){
-            throw new Exception404("해당 글들이 존재하지 않습니다");
-        }
+        Page<MentorPost> result = mentorSearchCategory.execute(keyword, pageable, mentorPostJPARepository);
+        List<MentorPostResponse.MentorPostAllDTO> mentorPostAllDTOS = result.stream().map(mentorPost -> {
+            List<UserInterest> writerInterests = userInterestJPARepository.findAllById(mentorPost.getWriter().getId());
+            return new MentorPostResponse.MentorPostAllDTO(mentorPost, writerInterests);
+        }).collect(Collectors.toList());
 
-        //List<MentorPost> mentorPostList = mentorPostJPARepostiory.findAll();
-        List<MentorPostResponse.MentorPostAllDTO> mentorPostDTOList = pageContent.getContent().stream().map(
-                mentorPost -> {
-                    List<UserInterest> writerInterests = userInterestJPARepository.findAllById(mentorPost.getWriter().getId());
-                    if(writerInterests.isEmpty()){
-                        throw new Exception404("해당 카테고리는 존재하지 않습니다");
-                    }
-
-                    return new MentorPostResponse.MentorPostAllDTO(mentorPost,writerInterests);
-                }
-        ).collect(Collectors.toList());
-        return mentorPostDTOList;
+        return mentorPostAllDTOS;
     }
 
     public MentorPostResponse.MentorPostDTO findMentorPost(int id){
@@ -71,6 +74,7 @@ public class MentorPostService {
                 .orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다.\n" + "id : " + id));
 
         //writer 데이터
+        //writer를 제외하고는 다 null 가능
         User mentor = mentorPost.getWriter();
         //mentee들 데이터
         List<NotConnectedRegisterUser> menteeList = contactJPARepository.findAllByMentorPostId(id);
@@ -81,53 +85,55 @@ public class MentorPostService {
                 .flatMap(mentee -> userInterestJPARepository.findAllById(mentee.getMenteeUser().getId()).stream())
                 .collect(Collectors.toList());
 
-        MentorPostResponse.MentorPostDTO mentorPostDTO = new MentorPostResponse.MentorPostDTO(mentorPost, mentorInterests, menteeList, menteeInterests);
-
-        return mentorPostDTO;
+        return new MentorPostResponse.MentorPostDTO(mentorPost, mentorInterests, menteeList, menteeInterests);
     }
     @Transactional
-    public void updateMentorPost(MentorPostRequest.CreateDTO createDTO, int id)
-    {
+    public void updateMentorPost(MentorPostRequest.CreateMentorPostDTO createMentorPostDTO, int id, User writer) {
+        isMentor(writer);
+
         MentorPost mentorPost = mentorPostJPARepository.findById(id).
                 orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));
 
-        try {
-            mentorPost.update(createDTO.getTitle(), createDTO.getContent());
-        } catch (Exception e) {
-            throw new Exception500("unknown server error");
-        }
+        validateUserAccess(writer, mentorPost);
+
+        mentorPost.update(createMentorPostDTO.getTitle(), createMentorPostDTO.getContent());
+
     }
 
-    public void deleteMentorPost(int id) {
-        try {
-            mentorPostJPARepository.deleteById(id);
-        } catch (Exception e) {
-            throw new Exception500("unknown server error");
-        }
+    public void deleteMentorPost(int id, User writer) {
+        isMentor(writer);
+
+        MentorPost mentorPost = mentorPostJPARepository.findById(id).
+                orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));
+
+        validateUserAccess(writer, mentorPost);
+
+        mentorPostJPARepository.deleteById(id);
     }
 
-    //생성 시간까지 조회하는 test service 코드 입니다
-    public List<MentorPostResponse.MentorPostAllWithTimeStampDTO> findAllMentorPostWithTimeStamp() {
-        List<MentorPost> pageContent = mentorPostJPARepository.findAll();
-        //List<MentorPost> mentorPostList = mentorPostJPARepostiory.findAll();
-        List<MentorPostResponse.MentorPostAllWithTimeStampDTO> mentorPostDTOList = pageContent.stream().map(
-                mentorPost -> {
-                    List<UserInterest> writerInterests = userInterestJPARepository.findAllById(mentorPost.getWriter().getId());
-                    if(writerInterests.isEmpty()){
-                        throw new Exception404("해당 카테고리는 존재하지 않습니다");
-                    }
+    public void changeMentorPostStatus(MentorPostRequest.StateDTO stateDTO, int id, User writer) {
 
-                    return new MentorPostResponse.MentorPostAllWithTimeStampDTO(mentorPost,writerInterests);
-                }
-        ).collect(Collectors.toList());
-        return mentorPostDTOList;
-    }
+        isMentor(writer);
 
-    public void changeMentorPostStatus(MentorPostRequest.StateDTO stateDTO, int id)
-    {
         MentorPost mentorPost = mentorPostJPARepository.findById(id)
-                .orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));;
-        mentorPost.changeStatus(stateDTO.getStateEnum());
+                .orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));
+
+        validateUserAccess(writer, mentorPost);
+
+        mentorPost.changeStatus(stateDTO.getMentorPostStateEnum());
+    }
+
+    private void isMentor(User writer) {
+        if ( writer.getRole() == Role.MENTEE ) {
+            throw new Exception401("권한이 없습니다");
+        }
+    }
+
+    private void validateUserAccess(User writer, MentorPost mentorPost) {
+        // 예외처리
+        if ( mentorPost.getWriter().getId() != writer.getId() ) {
+            throw new Exception401("권한이 없습니다.");
+        }
     }
 }
